fix(nxls): repair support for repos with only lerna.json
