+++ libs/shared/npm/src/index.ts
@@ -2,3 +2,4 @@ export * from './lib/workspace-dependencies';
 export * from './lib/find-nx-package-path';
 export * from './lib/package-details';
 export * from './lib/nx-json';
+export * from './lib/read-json';
